+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [4.11.0] - ?
 ### Added:
  - User *changePassword*, *deactivate* methods.
+ - AuthRequired exception.
 ### Changed:
  - User *login* method does not throw exception, now it returns *false* on fail.
  - Bot *getLastError* method returns string, not array.